stray coordinates on grid files change
