+++ test/test_subset.py
@@ -4,7 +4,6 @@ import uxarray as ux
 import uxarray.grid.slice as slice_module
 from uxarray.grid.slice import _remap_dense, _remap_kernel, _remap_searchsorted
 
-import numpy as np
 import pytest
 import xarray as xr
 
@@ -356,6 +355,38 @@ def test_bounding_box_with_stray_grid_time(gridpath):
     )
 
     # previously raised: "IndexError: dimension coordinate 'time' conflicts ..."
+    res = uxda.subset.bounding_box(lon_bounds=(-10, 10), lat_bounds=(-10, 10))
+
+    assert isinstance(res, ux.UxDataArray)
+    assert res.sizes["time"] == times.size
+    assert "n_face" in res.dims
+
+
+def test_bounding_box_with_stray_grid_time_dimension(gridpath):
+    """Subsetting must also work when the grid source carried a full ``time`` *dimension*
+    (a variable along ``time``) rather than the scalar coordinate of issue #1444.
+
+    A dimensional coordinate has dims ``(time,)``, which is not a subset of the
+    ``(n_face,)`` ``subgrid_face_indices`` indexer, so — unlike a 0-d coordinate — it
+    never attaches to that indexer and cannot collide during ``.isel``. The stray
+    dimension harmlessly rides along in ``Grid._ds`` but never reaches the indexing path.
+    """
+    # grid built from a source carrying a variable along a stray `time` dimension
+    grid_ds = xr.open_dataset(gridpath("ugrid", "quad-hexagon", "grid.nc"))
+    uxgrid = ux.open_grid(grid_ds.assign(stray=("time", [0.0, 1.0])))
+
+    # the bare dimension is not a coordinate, so it is not dropped; that is harmless
+    assert "time" in uxgrid._ds.dims
+
+    # a face-centered variable carrying its own time coordinate
+    times = np.array(["2018-01-01", "2020-01-01"], dtype="datetime64[ns]")
+    uxda = ux.UxDataArray(
+        data=np.ones((times.size, uxgrid.n_face)),
+        dims=("time", "n_face"),
+        coords={"time": times},
+        uxgrid=uxgrid,
+    )
+
     res = uxda.subset.bounding_box(lon_bounds=(-10, 10), lat_bounds=(-10, 10))
 
     assert isinstance(res, ux.UxDataArray)
